First file dump
This is a cleaned up copy of my private Nosta project repo. I removed private keys I had in places for testing (man, I hope I got them all), a bunch of crappy code (probably still some left).

There are still various incomplete features and details in here. But, a few people wanted the code to be in the open, so here we go.