Show stalls and products
Adds support for NIP-15. Stalls and products are now shown on a profile (that has the respective events). Clicking a product links out to the LNBits nostrmarket which has a proper shopping cart.

I hope that's enough new event types for now. Need to focus on tweaking all kinds of things next.