Adds a loader to the profile page while searching
It cycles through some emoji and copy, just for fun.

There's also functionality in there for deleting empty lists. Iris currently tracks when you checked your notifications or chats via empty lists, which creates a ton of spam events. This is a quick way to delete them. It's hidden because it was for testing and needs to be properly implemented.