Revision of the profile data overlay
Now it actually shows all the data that was loaded for the current profile. Since that's a lot, the nav has changed so you don't see all at once, but click through each content type.

What's still missing is intelligent evaluation and some options. Got some old profile versions or empty lists (looking at you, Iris) floating around? It should show you and offer a 1-click delete function. Also need to change the modal tag to "dialog", for better accessibility (just need to wrangle styling for that, it's a bit of mess).

So some detail work to do, but this should now be in good shape overall.