Show classifieds and foldering
The NIP for classifieds (402) is not merged yet, but it's so easy now to add new kinds with the existing structure, so why not.

Also organized the profile folder a bit better, moving files that belong to the same module into sub-folders ("products-*" > "products/*").

Calendar events (NIP-52) are also loaded now, but not shown in the main profile. You can take a look at them in the data overlay though. Another NIP that is not merged.

We'll see if adding unmerged NIPs is a good idea.