First implementation of user status
Implements NIP 38 for user statuses. Below the user name is now a small box showing the status. It can be just text, text with a link, or music. In the case of music, a musical note icon is also shown. A problem is that lots of music link go to the Spotify URI scheme. So if you don't have Spotify installed, the link doesn't work. Maybe clients who set statuses should use regular URLs instead?

If more status types are added, this functionality needs to be adjusted accordingly.