Add dynamic follow suggestions (#58)
* Add dynamic follow suggestions

The create flow has a screen where we suggest profiles to follow. In addition to the handful of static suggestions that's already been there, we now also load in trending profiles from the Nostr. band API.

Also sliding in a fix for user statuses. Don't show the component if there is a status with an empty String.

And adding labels and pinstr events to the data overlay (as preparation for showing them in the UI in a later update).

* Fixing icons

Not showing up in Safari without width and heights.