Adds profile tips
It's a small box you can see at the bottom of your own profile when logged in. It analyzes your Nostr activity and comes up with tips for things you haven't done yet. Each tip links out to nostrapp.link to see apps that support the respective kind. Hopefully this helps user discovery some new things to do.

This probably needs to be iterated over time to be really helpful. Some kinds are still missing, and it doesn't tell you to create notes or long-form posts because the site doesn't check for those.