Adding link helper class (#67)
This centralizes the creation of external links to other Nostr clients into a separate helper. The helper tries to see if the profile recommends a specific client for the note. If so, it uses that. If not, it uses a fallback.

Eventually, this could also consider the clients the logged-in user recommends.

The code is really messy. Just lots of looping over notes and tags. This can probably be cleaned up at some point and also be made more efficient (like identifying the preferred handler once when data comes in and not every time a link is generated). Step by step, as usual.