Update to lists handling (#70)
NIP 51 changed a lot recently. There are many more kinds to support (15 total), which are now all properly being shown on the site. For viewing lists, the site now links out to Listr, unless the profile has a different handler specified.

Can probably dig a bit deeper into what can be done with these lists. For example, there's an emoji list. Can that be used or shown in a fun way?