Add tracking (#77)
Adds the Umami tracking script. I don't like tracking, but the traffic has really increased a lot, so much so that Netlify is charging me more for the bandwidth. I need to figure out what people do on the site in order to optimize the experience (UX and bandwidth), and hopefully tracking will help with that. The data dashboard is publicly accessible. If it turns out to be not useful, or I don't end up looking at it regularly, I'll remove it again.

Umami is pretty good with privacy and is GDRP-compliant, so no cookie banners or disclaimers are needed.