Optimize image loading
The site currently prefetches all images, which causes a 10+ MB page load. It makes consecutive page views faster, but really is a waste and makes Netlify charge me for extra bandwidth. This is an unintended side-effect of using the useAssets composable. I don't 100% remember, but I think I used that composable because otherwise Vue does not include the respective images in the build. Could be wrong though.

This branch is a test to remove the composable, keep all images in the public folder, and load them from there. Going to test it with a preview build.