Fix for the header "/null" link issue
For whatever reason, the header button URL will sometimes not update to the logged-in users pubkey. I see this in the new Umami stats a good amount. Seems to be a problem with the client not updating from the server render, and also refs not being accessed by the value. Let's see if this fixes it.