Add follow/unfollow functionality (#85)
This enables a follow button on profiles. The majority of the work in this PR however is the session data handling that is happening in the background. In order to make a follow button work, the site needs to have the user's current follow list. To find that, it needs to have the user's relays. And those need to be gotten either from a browser extension, via NIP-05 handle, or other relays. And we don't want to use outdated lists either, or overwrite the users existing list with an almost empty one, in cases we don't find the former.

The sessionContactsService and sessionRelayService ideally handle all that stuff, but more testing is necessary to see how it holds up in various connectivity scenarios and profile configurations.

Once solid, those classes should enable further functionality, like showing mutual follows, updating relays, etc. But first, let's solidify this one.