Adds a "Republish" feature to the data overlay
The profile and follow list tabs in the data overlay now have subtle "Republish" text buttons (if you're logged in and the owner of that profile). Pressing will take that event and try to publish it to your relays.

With my own profile right now, publishing fails for most relays. Not cool. I need to check if those are dead by now, which seem to happen somewhat frequently. And there's an idea for another feature: helping users keep a clean relay list.

To keep users informed, I added a notification system. You'll get notifications slide in on the top right of the screen. They update as the publish status changes. They appear on success after 5s. Ones with errors stay around until you click them off.

So this is a helpful feature, but it's also a bit messy. Really needs more testing and polish.