Adds relay editing
Add editing of relays. The profile edit page now has a second tab for relays.

Just tested it and it works well. One problem I have is that the site gets its list of relays from Alby. But when post a new relay list event, Alby doesn't catch it right now. So the site still uses the old relay list. Not sure how to fix that yet.