Add a "Republish" feature to the data overlay (#90)
* Adds a "Republish" feature to the data overlay

The profile and follow list tabs in the data overlay now have subtle "Republish" text buttons (if you're logged in and the owner of that profile). Pressing will take that event and try to publish it to your relays.

With my own profile right now, publishing fails for most relays. Not cool. I need to check if those are dead by now, which seem to happen somewhat frequently. And there's an idea for another feature: helping users keep a clean relay list.

To keep users informed, I added a notification system. You'll get notifications slide in on the top right of the screen. They update as the publish status changes. They appear on success after 5s. Ones with errors stay around until you click them off.

So this is a helpful feature, but it's also a bit messy. Really needs more testing and polish.

* Adds relay editing

Add editing of relays. The profile edit page now has a second tab for relays.

Just tested it and it works well. One problem I have is that the site gets its list of relays from Alby. But when post a new relay list event, Alby doesn't catch it right now. So the site still uses the old relay list. Not sure how to fix that yet.

* Various fixes

* Some mobile layout fixes

* Making the relay list also republishable

* Various fixes