Nostr Tools update
Switching the library to be part of the code rather than a separate JS file.
Updated to the latest version of Nostr Tools, which required some changes.

Overall, it seems like relay connectivity is terrible now through this update. So need to properly test it and figure out what's up.